requirements.txt
astroid==3.1.0
blinker==1.7.0
click==8.1.7
dill==0.3.8
Flask==3.0.3
isort==5.13.2
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mccabe==0.7.0
platformdirs==4.2.0
pylint==3.1.0
python-dotenv==1.0.1
tomlkit==0.12.4
Werkzeug==3.0.2
